Bug Fix + Usability Fix
Error 1: Freeze Copy would cause the new object to be offset (not centered) relative to the object it was copied from.
Error 2: Object created from Freeze Copy would be selected after the copy was made instead of the original object remaining selected.
Error 3: Object created from Freeze Copy would be visible when it should be hidden.

Cause 1: Objects appear to be relatively offset when they are newly created.

Fix 1: Set the objects .location and .rotation_euler to be the same as the object it was copied from.