Bug Fix
Error: TypeError: object of type 'NoneType' has no len()

Cause: if any armature is hidden this error will be triggered. Even though the data shows the armature is correctly assigned to the variable, passing it to len() will throw the "noneType" error. Seems counter intuitive given the fact the armature exists, it's assigned to the variable, and the variable is clearly not None.

Fix: force checking for None type.