Added Decompile functionality.
The Compile feature copies the current object + armature, sets the object as a child of the current object, and then hides that copied armature and object. The decompile process simply takes those copies, unparents them, unhides them, deletes the current object, and then renames the extracted object to the name of the deleted object.