Bug Fix
Error: RuntimeError: Error: Armature 'Armature' not in edit mode, cannot add an editbone

Cause: After moving some code around it turned out that the bones were never selected before the Edit mode was toggled on inside the function that creates bones for the controller.

Fix: Forced the armature to be selected inside the function before edit mode is toggled.