Added function to retrive all the edges of a given face.
Added function to map an array of numbers in their low to high values.

Added a bevel function wrapper that ensures only the newly created faces are selected after the operation finishes as well as returning the selected faces.

Added function to check if an object with a given name exists.