Fixed invalid layer_map on non quad faces (Part 3)
Final bugs fixed. Layer Mapping works without issue. Had to rewrite generate_bevel_layer_map function to be not shit. This was the last one in need of rewriting to bring it up to the standards of the other mapping functions.

The only issue is it's slow as piss. I should optimize the speed some day but it's not important yet.